+++ client/src_initrd/busybox-0.60.5/libbb/libbb.h
@@ -200,7 +200,7 @@ enum {
 	MEGABYTE = (KILOBYTE*1024),
 	GIGABYTE = (MEGABYTE*1024)
 };
-const char *make_human_readable_str(unsigned long size, unsigned long block_size, unsigned long display_unit);
+const char *make_human_readable_str(unsigned long long size, unsigned long block_size, unsigned long display_unit);
 
 int ask_confirmation(void);
 int klogctl(int type, char * b, int len);
